Implement a practical solution for running the birds on the surface of a sphere
Runs the entire simulation in 3D coordinates, with the usual rules, but at each time step, normalizes the positions of the birds onto the sphere, and projects their velocities to be tangent to the sphere.

Full “winged” bird rendering in this new 3D version not yet complete.

Thanks to a group of mathematicians at the ETH for some insight into this one.